ask if unlink the download folder this is often convinient as usually it won't succeed at the first try

+++ tools/update_DT.R
@@ -286,4 +286,6 @@ local({
 })
 
 # clean up download folder
-unlink(dld_folder(), recursive = TRUE)
+if (isTRUE(askYesNo("unlink download folder?"))) {
+  unlink(dld_folder(), recursive = TRUE)
+}
